move around editor scripts and worked on Controller

+++ Assets/Scripts/Controllers/CharacterController.cs
@@ -63,7 +63,7 @@ public class CharacterController : FSMMonoBehaviour {
 	[Serializable]
 	public class JumpProperties
 	{
-		public float jumpHeight;
+		public int jumpHeight;
 		public bool canJump;
 		public int maxConsecutiveJumps;
 	}
